requirements.txt
# requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.109.2
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
pyperclip==1.8.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1